requirements.txt
torch
torchvision
numpy
PyYAML
pyarrow
matplotlib
imageio
Pillow
scikit-image
tqdm
wandb
